refactor: change did generator to service
